TODO: improve code quality, rename things,fail tolerance, stricter(implement) on incoming object validation,run in docker, try to reconnect,close the connection without server error, write unit test DONE: basic protos, implented 2 way communation with API, able to schedule pods based on resource

+++ ContainerOrchestrator.Base/Limitation.cs
@@ -2,7 +2,7 @@
 {
     public class Limitation
     {
-        public double CPU { get; set; }
+        public int CPU { get; set; }
         public int Memory { get; set; }
     }
 }
